Fix build by updating imports after move
